Add files via upload
Updated and created several versions of the original simulation function to be more adaptable to the new functions for demand generation, routing, etc. In addition, created a function which writes the location of each nurse over time into a CSV file to be used for visualization, and a function which optimizes the minimum number of nurses needed to keep customer wait times below 1 hour.